requirements.txt
requests==2.31.0
pyproj==3.6.1
langchain==0.0.348
urllib3==2.1.0
pandas==2.1.3
python-dotenv==1.0.0
openai==1.6.0
